Rename gameStart to waitingRoom
