+++ client/src/env.d.ts
@@ -2,7 +2,7 @@
 
 // https://vitejs.dev/guide/env-and-mode.html#intellisense-for-typescript
 interface ImportMetaEnv {
-    readonly VITE_DISABLE_GAMESTART: boolean
+    readonly VITE_DISABLE_WAITING_ROOM: boolean
     readonly VITE_BACKEND_URL: string
 }
