+++ client/src/index.ts
@@ -3,7 +3,7 @@ import { io, Socket } from "socket.io-client";
 
 import { SceneFullscreenScoreboard } from "./scene/SceneFullscreenScoreboard";
 import { SceneGameArena } from "./scene/SceneGameArena";
-import { SceneStartGame } from "./scene/SceneStartGame";
+import { SceneWaitingRoom } from "./scene/SceneWaitingRoom";
 import { GameState } from "./GameState";
 import { BOARD_SIZE, TILE_SIZE } from "common/shared";
 
@@ -17,9 +17,9 @@ const config = {
   parent: "root",
   width: BOARD_SIZE * TILE_SIZE,
   height: BOARD_SIZE * TILE_SIZE,
-  scene:  import.meta.env.VITE_DISABLE_GAMESTART
+  scene: import.meta.env.VITE_DISABLE_WAITING_ROOM
     ? [SceneGameArena, SceneFullscreenScoreboard]
-    : [SceneStartGame, SceneGameArena, SceneFullscreenScoreboard],
+    : [SceneWaitingRoom, SceneGameArena, SceneFullscreenScoreboard],
 };
 
 const socket = io(
@@ -30,8 +30,8 @@ const socket = io(
 const gameState = new GameState(socket);
 const game = new Phaser.Game(config);
 
-if (import.meta.env.VITE_DISABLE_GAMESTART) {
+if (import.meta.env.VITE_DISABLE_WAITING_ROOM) {
   game.scene.start("SceneGameArena", { gameState, socket });
 } else {
-  game.scene.start("SceneStartGame", { gameState, socket });
+  game.scene.start("SceneWaitingRoom", { gameState, socket });
 }
