+++ common/messages/sceneWaitingRoom.ts
